Thoroughly commented the codebase along with performing some minor refactors and bug fixes.
- Friend class is now ImmutableMappable instead of Mappable
- FileListVC allows deleting multiple files simultaneously
- FriendListVC does not show the option to scan QR Code on devices that do not have a camera
- FriendPickerVC selects the 'self' entry by default

+++ SwiftLock/ViewControllers/HomeViewController.swift
@@ -22,6 +22,7 @@ class HomeViewController: UITabBarController
                                                 FileManager.default.urls(for: .documentDirectory, in: .userDomainMask).first!.appendingPathComponent("Inbox")]
     }
 
+    // MARK: - View
     override func viewDidLoad() {
         super.viewDidLoad()
 
@@ -32,21 +33,56 @@ class HomeViewController: UITabBarController
     
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
+        
+        // segue to loginVC if no one's logged in
         if !CurrentUser.shared.isLoggedIn {
             performSegue(withIdentifier: SegueIds.ToLogin, sender: self)
         }
     }
     
-    @objc fileprivate func userLoggedIn() {
-        if let _ = presentedViewController as? LoginViewController {
-            dismiss(animated: true, completion: nil)
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // set self as the delegate of FriendPickerViewController
+        if segue.identifier == SegueIds.ToFriendPicker,
+            let friendPicker = segue.destination.mainVC as? FriendPickerViewController {
+            friendPicker.delegate = self
         }
     }
     
-    @objc fileprivate func userLoggedOut() {
-        performSegue(withIdentifier: SegueIds.ToLogin, sender: self)
+    // The progressHUD to display during (en/de)cryption process
+    fileprivate var progressHUD: MBProgressHUD? {
+        didSet {
+            // setup the progressHUD
+            progressHUD?.removeFromSuperViewOnHide = true
+            progressHUD?.mode = .annularDeterminate
+        }
     }
     
+    // MARK: - Login/Logout event handling
+    
+    /// dimisses the presented VC, if it's a LoginVC
+    @objc fileprivate func userLoggedIn() {
+        // dispatch off to main q
+        DispatchQueue.main.async { [weak self] in
+            if self?.presentedViewController is LoginViewController {
+                self?.dismiss(animated: true, completion: nil)
+            }
+        }
+    }
+    
+    /// segues to the loginVC
+    @objc fileprivate func userLoggedOut() {
+        // dispatch off to main q
+        DispatchQueue.main.async { [weak self] in
+            self?.performSegue(withIdentifier: SegueIds.ToLogin, sender: self)
+        }
+    }
+    
+    // MARK: - Encryption/Decryption
+
+    /// If 'currentFile' is presently nil, sets it to 'url'. Otherwise __deletes__ the file at 'url'
+    /// Call this method to encrypt/decrypt a file
+    ///
+    /// - Parameter url: URL of the file to encrypt/decrypt
     func handleFile(url: URL) {
         guard CurrentUser.shared.isLoggedIn else {
             presentedViewController?.alert(withTitle: Strings.TryingToOpenFileWhenLoggedOut, message: nil)
@@ -57,7 +93,9 @@ class HomeViewController: UITabBarController
             currentFile = url
         } else {
             do {
-                try FileManager.default.removeItem(at: url)
+                if url.isFileInTemporaryLocation {
+                    try FileManager.default.removeItem(at: url)
+                }
             } catch (let error) {
                 print("Error deleting the source file: ", error)
             }
@@ -66,12 +104,15 @@ class HomeViewController: UITabBarController
         }
     }
 
+    /// Holds the URL to the file currently being encrypted/decrypted
+    /// Setting this to a non-nil value starts the process of (en/de)cryption.
     fileprivate var currentFile: URL? {
         didSet {
             guard let url = currentFile  else  {
                 return
             }
             
+            // check whether file is encrypted or not
             let fileIsEncrypted: Bool
             do {
                 fileIsEncrypted = try MiniLock.isEncryptedFile(url: url)
@@ -81,13 +122,15 @@ class HomeViewController: UITabBarController
             }
             
             if fileIsEncrypted {
+                // decrypt the encrypted file
                 decrypt(url)
             } else {
+                // wait for the dismissal of a presented VC (typically a share sheet) using a semaphore
                 let semaphore = DispatchSemaphore(value: 0)
-                // if a vc is on top, dismiss it first
                 if presentedViewController != nil {
                     presentedViewController?.dismiss(animated: true) { semaphore.signal() }
                 } else {
+                    // no VC on top, we are good to go!
                     semaphore.signal()
                 }
                 
@@ -104,29 +147,24 @@ class HomeViewController: UITabBarController
         }
     }
     
-    fileprivate var progressHUD: MBProgressHUD? {
-        didSet {
-            progressHUD?.removeFromSuperViewOnHide = true
-            progressHUD?.mode = .annularDeterminate
-        }
-    }
-    
+    /// Starts decrypting the file at 'url'
+    ///
+    /// - Parameter url: URL of the file to decrypt
     fileprivate func decrypt(_ url: URL) {
-        print("Decrypting...")
-        
+        // setup a progressHUD
         progressHUD = MBProgressHUD.showAdded(to: self.view, animated: true)
         progressHUD!.label.text = Strings.DecryptingMessageInProgressHUD
         
-        var decryptor: MiniLock.FileDecryptor
+        // initialize and configure a MiniLock.FileDecryptor object
+        let decryptor: MiniLock.FileDecryptor
         do {
             decryptor = try MiniLock.FileDecryptor(sourceFile: url, recipientKeys: CurrentUser.shared.keyPair!)
         } catch {
-            DispatchQueue.main.async { [weak self] in
-                self?.progressHUD?.hide(animated: true)
-                self?.alert(withTitle: Strings.ErrorDecryptingFile, message: error as? String)
-                self?.currentFile = nil
-            }
-            
+            // handle init errors
+            progressHUD?.hide(animated: true)
+            alert(withTitle: Strings.ErrorDecryptingFile, message: error as? String)
+            currentFile = nil
+
             return
         }
         
@@ -134,16 +172,19 @@ class HomeViewController: UITabBarController
         
         // go off the main q
         DispatchQueue.global(qos: .userInitiated).async { [weak self] in
+            // set currentFile to nil before returning
             defer {
                 self?.currentFile = nil
             }
 
-            var decryptedFile: URL
+            // start decrypting the file
+            let decryptedFile: URL
             do {
                  decryptedFile = try decryptor.decrypt(destinationDirectory: CurrentUser.shared.decryptedDir,
                                                        filename: nil,
-                                                       deleteSourceFile: HomeViewController.isFileInTemporaryLocation(url: url))
+                                                       deleteSourceFile: url.isFileInTemporaryLocation)
             } catch {
+                // hide progressHUD and alert the user
                 DispatchQueue.main.async {
                     self?.progressHUD?.hide(animated: true)
                     self?.alert(withTitle: Strings.ErrorDecryptingFile, message: error as? String)
@@ -159,6 +200,8 @@ class HomeViewController: UITabBarController
                 
                 strongSelf.progressHUD?.hide(animated: true)
                 strongSelf.fileListVC?.tableView?.reloadData()          // updates the file size in FileListVC
+                
+                // show a share sheet for the file that was just decrypted
                 let activityVC = UIActivityViewController(activityItems: [decryptedFile], applicationActivities: nil)
                 activityVC.modalPresentationStyle = .popover
                 
@@ -167,21 +210,27 @@ class HomeViewController: UITabBarController
         }
     }
     
+    /// Starts encrypting the file at 'url'
+    ///
+    /// - Parameters:
+    ///   - url: URL of the file to encrypt
+    ///   - friends: recipients to whom the file should be encrypted
     fileprivate func encrypt(_ url: URL, to friends: [Friend]) {
+        // setup a progressHUD
         progressHUD = MBProgressHUD.showAdded(to: self.view, animated: true)
         progressHUD!.label.text = Strings.EncryptingMessageInProgressHUD
 
-        var encryptor: MiniLock.FileEncryptor
+        // initialize and setup a MiniLock.FileEncryptor object
+        let encryptor: MiniLock.FileEncryptor
         do {
             encryptor = try MiniLock.FileEncryptor(fileURL: url,
                                                    sender: CurrentUser.shared.keyPair!,
                                                    recipients: friends.map { $0.id } )
         } catch {
-            DispatchQueue.main.async { [weak self] in
-                self?.progressHUD?.hide(animated: true)
-                self?.alert(withTitle: Strings.ErrorEncryptingFile, message: error as? String)
-                self?.currentFile = nil
-            }
+            // handle init errors
+            progressHUD?.hide(animated: true)
+            alert(withTitle: Strings.ErrorEncryptingFile, message: error as? String)
+            currentFile = nil
 
             return
         }
@@ -190,17 +239,21 @@ class HomeViewController: UITabBarController
         
         // go off the main q
         DispatchQueue.global(qos: .userInitiated).async { [weak self] in
+            // set currentFile to nil before returning
             defer {
                 self?.currentFile = nil
             }
 
-            var encryptedFile: URL
+            // start encrypting the file
+            let encryptedFile: URL
             do {
                 encryptedFile = try encryptor.encrypt(destinationDirectory: CurrentUser.shared.encryptedDir,
                                                       filename: nil,
-                                                      deleteSourceFile: HomeViewController.isFileInTemporaryLocation(url: url))
+                                                      deleteSourceFile: url.isFileInTemporaryLocation)
             } catch {
+                // dispatch back to the main Q
                 DispatchQueue.main.async {
+                    // hide progressHUD and alert the user
                     self?.progressHUD?.hide(animated: true)
                     self?.alert(withTitle: Strings.ErrorEncryptingFile, message: error as? String)
                 }
@@ -216,6 +269,8 @@ class HomeViewController: UITabBarController
 
                 strongSelf.progressHUD?.hide(animated: true)
                 strongSelf.fileListVC?.tableView?.reloadData()          // updates the file size in FileListVC
+                
+                // show a share sheet for the file that was just encrypted
                 let activityVC = UIActivityViewController(activityItems: [encryptedFile], applicationActivities: nil)
                 activityVC.modalPresentationStyle = .popover
                 
@@ -224,39 +279,27 @@ class HomeViewController: UITabBarController
         }
     }
     
-    fileprivate class func isFileInTemporaryLocation(url: URL) -> Bool {
-        let parentDir = url.deletingLastPathComponent()
-        return Constants.TemporaryLocations.contains(parentDir)
-    }
-    
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        if segue.identifier == SegueIds.ToFriendPicker,
-            let friendPicker = segue.destination.mainVC as? FriendPickerViewController {
-            friendPicker.delegate = self
-        }
-    }
-    
+    /// This will be the first SLFileListViewController in self's list of vcs
+    /// In case there isn't even one of those, its value will be nil
     lazy var fileListVC: SLFileListViewController? = {
         if let viewControllers = self.viewControllers {
             for vc in viewControllers {
                 if let fileListVC = vc.mainVC as? SLFileListViewController {
-                    print("Found fileList!")
                     return fileListVC
                 }
             }
         }
         
-        print("FileListVC not found!")
         return nil
     }()
 }
 
+// MARK: - FriendPickerDelegate
 extension HomeViewController: FriendPickerDelegate {
     func friendPicker(_ picker: FriendPickerViewController, didPickFriends friends: [Friend]) {
-        if presentedViewController?.mainVC == picker {
-            dismiss(animated: true, completion: nil)
-        }
-
+        dismiss(animated: true, completion: nil)
+        
+        // encrypt 'currentFile' with 'friends' as the recipients
         guard let url = currentFile else {
             return
         }
@@ -265,17 +308,34 @@ extension HomeViewController: FriendPickerDelegate {
     }
     
     func friendPickerDidCancel(_ picker: FriendPickerViewController) {
-        if presentedViewController?.mainVC == picker {
-            dismiss(animated: true, completion: nil)
+        dismiss(animated: true, completion: nil)
+
+        if let url = currentFile {
+            // delete currentFile if it is in a temp directory
+            if url.isFileInTemporaryLocation {
+                try? FileManager.default.removeItem(at: url)
+            }
+
             currentFile = nil
         }
     }
 }
 
+// MARK: - MiniLockProcessDelegate
 extension HomeViewController: MiniLockProcessDelegate {
     func setProgress(to progress: Double, process: MiniLockProcess) {
+        // update the progressHUD
         DispatchQueue.main.async { [weak self] in
             self?.progressHUD?.progress = Float(progress)
         }
+    }
+}
+
+// MARK: - Misc.
+extension URL {
+    /// true if the url's direct parent is one of the directories `HomeViewController.Constants.TemporaryLocations`
+    var isFileInTemporaryLocation: Bool {
+        let parentDir = self.deletingLastPathComponent()
+        return HomeViewController.Constants.TemporaryLocations.contains(parentDir)
     }
 }
